requirements.txt
playwright==1.49.1
pytest==8.3.4
pytest-playwright==0.6.2
allure-pytest==2.13.5

